+++ src/program/movie/MovieActionPaint.h
@@ -31,6 +31,7 @@ class MovieFileInputs;
 class MovieActionPaint : public IMovieAction {
 public:
     MovieActionPaint(uint64_t start_frame, uint64_t end_frame, SingleInput si, int newV, MovieFileInputs* movie_inputs);
+    MovieActionPaint(uint64_t start_frame, SingleInput si, const std::vector<int>& newV, MovieFileInputs* movie_inputs);
 
     void undo(MovieFileInputs* movie_inputs);
     void redo(MovieFileInputs* movie_inputs);
@@ -38,6 +39,7 @@ public:
 private:
     SingleInput input;
     std::vector<int> old_values;
+    std::vector<int> new_values;
     int new_value;
 };